refactored out helpers from post/put listing routes
